Move SlotHooks to a separate file
